refactor(gemini): relocate steane_defaults from lanes to gemini
