+++ python/bloqade/gemini/steane_defaults.py
