+++ python/bloqade/lanes/__init__.py
@@ -11,7 +11,3 @@ from .noise_model import (
     generate_simple_noise_model as generate_simple_noise_model,
 )
 from .rewrite.move2squin.noise import NoiseModelABC as NoiseModelABC
-from .steane_defaults import (
-    steane7_m2dets as steane7_m2dets,
-    steane7_m2obs as steane7_m2obs,
-)
